feat: add middleware and change readme
